Don't initialize Aphid.UI.View instances with a viewName.
This puts us in a state where we must initialize from an element or from a subclass. This API needs to change to allow options to be passed for initialization, or to use a setter pattern so that we can load views on demand. I am undecided in which approach to take at this time.